:zap: fix: cors config 수정
