--- src/main/java/com/codez4/meetfolio/global/config/WebConfig.java
@@ -1,26 +1,0 @@
-package com.codez4.meetfolio.global.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpMethod;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebConfig implements WebMvcConfigurer {
-
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**")
-            .allowedOrigins("http://34.64.177.41:3000", "http://www.meetfolio.kro.kr:3000", "http://34.64.177.41:9090")
-            .allowedMethods(
-                HttpMethod.GET.name(),
-                HttpMethod.POST.name(),
-                HttpMethod.PUT.name(),
-                HttpMethod.DELETE.name(),
-                HttpMethod.PATCH.name())
-            .allowCredentials(false)
-            .maxAge(3600)
-            .exposedHeaders("Authorization", "RefreshToken");
-    }
-
-}
